+++ packages/vike-react-query/src/+config.h.ts
@@ -1,5 +1,3 @@
-import type { QueryClientConfig } from "@tanstack/react-query";
-import type { ReactNode } from "react";
 import type { Config } from "vike/types";
 
 const config: Config = {
@@ -22,14 +20,5 @@ const config: Config = {
     },
   },
 };
-
-declare global {
-  namespace VikePackages {
-    interface ConfigVikeReact {
-      queryClientConfig?: QueryClientConfig;
-      FallbackErrorBoundary: ((props: { children: ReactNode }) => ReactNode) | string;
-    }
-  }
-}
 
 export default config;