+++ packages/vike-react-query/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./WithFallback";
+export * from "./typing.d";