+++ packages/vike-react-query/src/typing.d.ts
@@ -1,0 +1,13 @@
+import type { QueryClientConfig } from "@tanstack/react-query";
+import type { ReactNode } from "react";
+
+declare global {
+  namespace VikePackages {
+    interface ConfigVikeReact {
+      queryClientConfig?: QueryClientConfig;
+      FallbackErrorBoundary: ((props: { children: ReactNode }) => ReactNode) | string;
+    }
+  }
+}
+
+export {}