+++ packages/vike-react/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./ClientOnly";
+export * from "./typing.d";