+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @techmely/cache
 
+## 1.0.1
+
+### Patch Changes
+
+- Upgrade deps
+
 ## 1.0.0
 
 ### Major Changes