+++ packages/nuxt-partytown/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @techmely/nuxt-partytown
 
+## 1.0.1
+
+### Patch Changes
+
+- Upgrade deps
+
 ## 1.0.0
 
 ### Major Changes